Separated client stats container to own component
